Create user models required by next-auth model. (#22)
This schema is allow us to use next-auth's prisma adapter in the frontend out of box. We can customise this model later.

A User has one or more accounts (logging in with Google is an account; using another oauth provider creates another account, but can be linked to the same user).

Currently email is required since we always use oauth.

fixes YUP-142