Store model streaming metrics in the chat messages table for UI display. (#175)
We want to see how long each individual messages take to stream, so they need to be persisted in a relational db. This column should not be used to store monitoring / operational metrics data, though. It should only be used to drive the user interface.

Here is how the use case looks like

ref YUP-464