Use global variable instead of contextvar while creating async engine. (#313)
Creating engine in a context var adds a significant latency, wasted in initialization.

Latency difference between sync and async with ContextVar is ~100ms+ (see image: blue is async, green is sync). Need to verify if this is reduced with a global var.

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/ZNNpznpLhnMzanob0R9n/eb29e0bb-672b-4753-9727-8016a77a192b.png)

Engine is expected to be a global variable by SQLAlchemy and it will not leak information between requests.

cc: @daemon