added retries and post to slack incase of exception for claim reward (#468)
added retries and post to slack incase of exception for claim reward.

Incase the claim reward fails for either feedback or QT eval, it will retry and then even after the retries if it fails, then it will log an exception and do a slack post.

Post kabini, all these direct credits would anyway move to scratchcard based credits in which case, we don't have to claim credits right away. So this is a temp gap to handle occasional failure incase of claiming credits