ENG-1092 Implement static credits for QT & FB (#477)
Implement static credit for QT & Feedback

The latency introduced with the call to the backend is causing an experience issue. The idea is that we will issue a static credit on the frontend and pass that amount to the backend amount to process it. This is the backend change to support taking the frontend decided amount as an input. 

To avoid users gaming the system, the backend has checks around bounds. It's possible that this backend claim might fail in which case a slack post will notify us. Not the best way to build but would work ATM. 

Long term would be to build a queue based system.