for QT_eval pass message_id for reward (#481)
for QT_eval pass message_id for reward

This is because the turn_id is not available in the frontend Thumbsup component. However message_id is. So message_id is passed to the backend for QT_Eval which is used to retrieve the turn_id. Since this is all happening async, an additional query doesn't have any latency issues