pass feedback comment to backend for reward calculation (#496)
pass feedback comment to backend for reward calculation.

currently we are just using the length of the feedback to change the mean. This way higher feedback content would get slightly higher reward. however in the long run, we would call an LLM model to judge the quality and reward accordingly.