updated config.py to address pydantic related test issues coming in staging (#540)
after upgrading pydantic to 2.10 (which is needed by coinbase sdk), our tests are failing in staging as the db related values are blank leading to a blank db url. This is a new check in pydantic 2.10. So updated config with some default values to ensure pytest clears in staging. Also added a new method to ensure that in non local env, these values are required to avoid a situation of these variables getting used in prod