created an entry in the payment_instruments for crypto (#542)
The payment_instruments table holds all the different payment instruments present in the system including that of our own bank account or wallet addresses.

This table is needed as it's PK is used as a foreign key in the payment_transactions table.

This is loading the value from an env variable so that we can provide a different value when we do the prod run. Even if we insert prod with some wallet address, it doesn't matter as the wallet really gets loaded from the google cloud file and we can update the address later.