do not raise error incase wallet initialization fails (#555)
do not raise error incase wallet initialization fails

Wallet loading usually takes time. So it's better to load it as part of server startup so that when the 1st request hits, it doesn't take a long time.

However the current code was throwing error incase wallet initialization doesn't happen. This prevents the whole server from starting up. So this change doesn't raise that error and logs an exception. Incase the wallet is not there at the time of startup, the first request will try to load it