add an enum inference_failed to language model response status table (#607)
add an enum inference_failed to language model response status table.

Idea is to make an entry in this table with this status for a model, incase the inference job fails. If 2 consecutive cron job fails with this status then inactivate the model. This PR is just adding an enum