added diff facilitators as well as taking facilitator type as input for facilitator init (#619)
added diff facilitators as well as taking facilitator type as input for facilitator init.

For a crypto destination, look for different types of facilitators and incase none of them is passed then use the onchain facilitator. This will ensure the current integration doesn't break incase a facilitator type is not passed from frontend for a crypto destination