add cbbtc as an enum and enable gasless for this (#644)
add cbbtc as an enum and enable gasless for this.

It seems coinbase can do gasless for cbbtc too. At some point we might be looking for BTC too. Hence added a new enum as well as enabled gasless for this.

TODO - At some point, we might want to create a type to list the tokens supported by gasless and remove the or condition but for the time being this is simple