added an enum of reversed to indicate the records that were created against failed txns (#657)
added an enum of reversed to indicate the records that were created against failed txns

Today when a transaction fails, we create a reversal row in payment transactions against the failed txn. However the status of that record is marked as success. That could be confusing against other successful txns. Hence creating an explicit status to indicate the status as reversed for the transactions that were created for this purpose. This should help in reconciliation better