ENG-1471 add user name to slack post (#717)
ENG-1471 add user name to slack post.

Amplitude had a method to retrieve a list of user names given a user id. Refactored that method to a utils class and added another method to get name for a specific user.

I could have reused the earlier method by passing a single ID but the in query is inefficient if we know we are looking for 1 user. Hence a diff method