Implement "verify VPA" as validate-destination-identifier. (#782)
Made it generic in case we have other cases where we want to validate the user's input.

The output of the validation will contain a short lived encrypted token that contains the validated destination.

Why do we need to call VerifyVPA before payment?
If a user enters a phone number, the API gives us the corresponding VPA associated with that phone number. We're expected to use that VPA to make the UPI payment request to the bank.
Since we anyway want to allow the user to validate the VPA before submitting the request, we can get away with making the call only once. Encrypting the response ensures that the client doesn't know of the resolved VPA for the phone number, thus preserving privacy for abusers who want to call this API just to know who owns the phone number.