change log level to warning (#792)
change log level to warning when the async polling times out from CB.

Today when the CB polling times out, we are logging an error. CB has confirmed that they have an issue on their side. These time outs are frequent and triggering an alert on our side.

Changing this to warning as anyway we have a cron job that does a polling every hour for pending txns.