Mock Axis UAT responses in nonprod envs. (#862)
Axis UAT env is pretty flaky, and their servers return different responses than what we'd generally expect. We will try to get the situation to be more stable, but we don't expect too much from the bank.

We plan to mock the responses in the nonprod environment, if Axis returns a different response than what we expect. 
We will still test the network and gateway and ensure that mTLS, credentials and encryption is still working as intended. Axis' UAT network and gateway seems to be pretty stable, so we're not  (mostly) mocking this out.

We're also introducing magic phone numbers which will give the response that we need to test the Verify VPA flow.

ref ENG-609