Allow direct network access to the default VPC. (#870)
ref ENG-609

Axis needs us to connect from certain static IPs, requires mTLS along with application layer encryption.
We found that the best setup is to have a set of proxy machines that yupp-mind talks to from anywhere. These proxies are thin nginx servers with static outbound IP addresses that do mTLS too.

To connect to these proxies from yupp-mind, we use the existing VPC and use the internal IP addresses of those proxies in the URL