Use the validated_vpa from the instrument_metadata instead of the destination_identifier. (#877)
The client sends the phone number as the destination_identifier, but we should use the validated_vpa instead, which we received from the bank in the VerifyVPA call for that phone number.

Confirmed with the bank that they do not support phone number as the account number directly, we need to translate the phone number to a VPA in the previous step.