Send user visible error messages for the cashout flow from the server. (#932)
Cashout has a lot of validations at multiple places. Doing a first pass at tagging whatever could be user visible messages to display it to the user.

All ValueError messages are sent as HTTP 400 errors, and any explicit HTTP exceptions are sent to the yupp-head server, which will use the error strings in the 400 response to show it to the user.

ref ENG-1671, ENG-1591, ENG-1685 and others.