create user capability override table (#941)
create user capability override table

As we roll out new capabilities, we might want to restrict the whole capability or certain aspects of it to specific users. By default, all users would have access to all capabilities but if we want to override then we will make entries here.

For Ex:- if we want to disable cashout activity to some users while letting them continue to provide model feedback, we can make an entry in this table for the user and the cashout capability with status as disabled