create a new events table to keep track of different events (#964)
create a new events table to keep track of different events

This table will help in keeping track of the different events happening in both frontend and backend. This will help when someone is trying to look at a user's activity in soul. We also don't need to scrape Slack to see activity details.

Long run, this table might grow at the fastest rate and needs to be looked into from a partition/sharding etc. perspective. Or we could simply prune the table of older records once we have a data warehouse