handle errors from CB to propagate to client (#971)
handle errors from CB to propagate to client. Raise ValueError so that error message from CB is passed to frontend to be shown on client.

An alternate option is to continue to show generic error but this doesn't give the user an escape hatch to self serve. For Ex:- if the email is restricted on coinbase side, then the user wouldn't realize that