Add a model speed reranker for models after final topK (#994)
So the faster models is always on the left. This is based on the speed information inferred from chat_message, similar to how we dashboarded it in superset, we use p50 and p90 value of time-to-first-token and token-per-sec to estimate their 95% confidence interval and use their respective upper/lower bound to estimate a conservative overall latency, with some tunable weights (depending on if we care about TTFT or overall speed), and normalize into a 0-1 score.