weekly job to check balanced ledger (#1048)
Weekly job to ensure there is no mismatched balance.

point_transactions is our ledger. However today there is a mismatch between what is there in the ledger vs. points present in the user table. This job is to validate on a weekly basis how many such users are there so that we can start addressing them.

Long term - This should move to run off a read replicate or data warehouse. However this should help in the short term