create a user vendor profile table (#1061)
create a user vendor profile table.

As we interact with different vendors, vendors might want to store our users on their side and would require us to pass that ID for future interactions related to this user. So created a field for that and added an additional columns incase we want to capture anything else. 

Had forgotten to add a vendor name. Added that in a later PR