Add QuickTake fallback ability (#1075)
Add a new for_fallback mode for quicktake, so it can generate an even faster and contextual answer as a fallback, but not necessarily fully answer the question. This should cover many <CANT_ANSWER> or timeout case from the original quicktake. It's only meant to be used as a fallback but not the first choice.