create a new enum for partner identifier (#1082)
create a new enum for partner identifier

For some of the partners, we might not store the end destination account details and the partner will store them. The partner will send us a token mapping to that which we can store. In that case the identifier type will not work for any of the current ones and hence created a new enum for partner identifier