check for repeat activity (#1222)
check for repeat activity

As part of event creation, check if events are firing for more than the number of times for the same user over a pre-configured period and log to slack incase it happens. There is a killswitch incase we want to turn it off.

This won't have any user impact as the create event returns first and then tries to create the event.