INR follows a diff conversion rate and this is to fix that (#1284)
INR follows a different conversion rate than other currencies. It has a special rate. So the usd amount calculation is not right. The yupp to USD rate shouldn't be used here. Instead based on the INR amount, we need to find the USD amount.

Had a nightmare that I might have screwed up earlier code by making an assumption. Woke up to find it was actually true!!!