support default network for other currencies (#1354)
support default network for other currencies

Why - We soon have to support other assets. For assets that are supported via base, we want to use that route as it's less expensive. However incase an asset is not available on base like BTC, we should use the default network. Here is a snapshot of coinbase doc. So we specify the network only if it is ETH or USDC for now (may be TRUMP later)

![Screenshot 2025-02-12 at 12.53.46 PM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/YnvSmrL3M7Guz5YA8vwp/20e7fc37-82fe-4e1a-a367-519e367ca0c0.png)