Test a new deployment workflow (#1368)
The attempt is to build the image once and deploy it for each service.

This means that the env file is created per service, and it can't part of the built image, and it is created when we deploy that service.

TODO is to allow multiple deployments in one workflow. The current PR just tests if the backend deployment works as intended without issues.

This PR needs to be submitted for me to test further.