Fix the v2 deployment script (#1382)
This is a step towards build once, deploy multiple times.

This change removes the create-env yaml, and merges it back, similar to how it was earlier as GitHub doesn't allow splitting workflow files, and also sharing the secrets with that new file.

It also creates a .env.yaml file which is sent to gcloud run deploy. The Settings class in config.py will now load the env variables from the OS env, instead of from the .env file.
Note that postgres' password had to be escaped to work with quotes.

Tested this with backend-staging and partner-payments-server-staging. 

Also updated the schedule to run 1 hour after the regular/previous deploy, just to verify if things are working fine.

TODO:
- Support webhooks
- Split Build and Deploy into two steps
- Allow deploying multiple images at once