add column to capture reversed txn id incase this is an adjustment (#1457)
add column to capture reversed txn id incase this is an adjustment.

Use Case - Sometime bad actors give us junk app feedback to earn rewards. We might want to reverse some of these rewards and adjust their points. When such an adjustment happens, we want to capture the txn id that is being reversed. Hence the need for a nullable column