ENG-2371 show if cashout is disabled for some user (#1502)
ENG-2371 show if cashout is disabled for some user on soul. So we need to return the value incase cashout has been disabled for some user. By default everyone has access. So this needs a bit of a complicated join to identify

As part of this, moved the get_users method from the router to the service class. Also refactored the class to make it a little easier to understand the different order in which queries are fired and moved reusable code to common methods.