Convert Claude thinking chunks to flat text compatible with R1's output (#1524)
When extended-thinking is enabled in Claude, it returns json chunks rather than text chunks. The "type" field indicates if it is a thinking chunk or a text chunk. See doc string for `_process_claude_thinking_content()` in the PR for examples.

Output is compatible with R1's thinking output. For FE both look the same. E.g.
```
<think>

Bunch of thoughts...

</think>

Response text to user...
```

Ref ENG-2398