enable retry for coinbase.com payout attempts (#1630)
enable retry for coinbase.com payout attempts

Details - Often the coinbase.com API call fails even though coinbase actually processes the transaction on their side. This leaves the txns in a pending state on our system and since we never get a tx id, we can't even find out the status of the tx.

Enabled a retry in such exceptions. Since we have an idem field, hence in this case Coinbase will not do another transfer and instead return the prior one per their documentation. The idem is the payment txn id and hence it's safe to retry

![Screenshot 2025-03-04 at 1.25.24 PM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/YnvSmrL3M7Guz5YA8vwp/d335f78e-c64a-4ee1-836c-5a89be8134df.png)