stripe integration - add github token (#1668)
stripe integration - add github token

Stripe's SDK is a private repo and not public. They have only enabled access to my GH ID. We need to generate a token from that ID to be able to access the SDK. I've stored the token as a GH secret so that we can access it during deployment.