refactor payment link code (#1783)
This is a refactoring. 

Earlier we had the code under the user route but the more I think the more it feels that it should be under payments. As part of this we need to return the payment link which the user would be redirected to. 

Have not removed the old code as front end changes are still required. Would remove that after FE changes go in and are integrated