Youtube : Use second labeler for more fault tolerance (#1805)
This PR adds `gpt-4o-mini` in addition to `llama-3.3-70b-versatile` (on Groq) for labelling prompt for youtube videos. 
This it to handle cases like today where `llama-3.3-70` was timing out for many queries. 

We take which ever replies first. Updated logging to more easily see which model returned the response first.

Tested manually.